use new admin client
